oci_discovery/ref_engine_discovery: Catch HTTPError for ref_engine.resolve
Instead of catching URLError.  URLError is for protocol-level issues
(e.g. wrapping an SSLError if you attempt an HTTPS connection to a
server that does not speak TLS).  They do not have much structure [1].

HTTPError, on the other hand, is for things like 400 errors [2] and
they do support things like geturl [2,3].

[1]: https://docs.python.org/3/library/urllib.error.html#urllib.error.URLError
[2]: https://docs.python.org/3/library/urllib.error.html#urllib.error.HTTPError
[3]: https://docs.python.org/3/library/urllib.request.html#urllib.request.urlopen