oci_discovery/uri_template: Stub for URI Template expansion
Python's .format can handle things like {host}, but there's a lot more
than that in RFC 6570.  Instead of rolling our own implementation,
punt to the uritemplate package [1].  The license is [1,2]:

   BSD-3-Clause OR Apache-2.0

which should be compatible with the OCI's Apache-2.0 requirement [3],
even if that requirement applies to third-party dependencies, which is
not clear to me.

[1]: https://pypi.python.org/pypi/uritemplate
[2]: https://github.com/python-hyper/uritemplate/blob/3.0.0/LICENSE
[3]: https://www.opencontainers.org/about/governance
     Section 8.a