tools/refengine: Add a ref-engine protocol registry
And add ref-engine references.  This cleans up the ref-engine portion
of the previous tools/ implementation.  I have tried to touch the
ref-engine discovery portion as little as possible, although there's
some cleanup work still to do there as well.

The newimagespec package has the image-spec/specs-go/v1 updates that
I'd like to see upstreamed (a new casEngines descriptor property, and
descriptor consumers that use it).  If/when those eventually land in
image-spec we'll deprecate/drop the newimagespec package here.

I've removed the tools/ prefix from the tools/README.md example,
because that README is already inside the tools directory.