tools/refenginediscovery: Refactor Ref-Engine Discovery library API
Pull the UI portions of this out into tools/cmd, because the library
API should not depend on urfave/cli.

The new refenginediscovery API is based on a per-configured-ref-engine
callback, which gives callers lots of flexibility in how they want to
handle each ref-engine and if/when they want to abort the
protocol/host iteration.

I've also adjusted the ref-engine interface to return a new MerkleRoot
type.  This improves on the old Descriptor returns by:

* Being less opinionated about the root object, to stay in keeping
  with ref-engine-discovery.md's:

    The Merkle root may be a descriptor (media type
    application/vnd.oci.descriptor.v1+json), or it may have a
    different media type.

* Including the retrieval URI, which may be needed to expand
  references in the root object (e.g. if the root has a casEngines
  configuration which includes a relative reference).

These changes bring the API and output of the 'oci-discovery resolve
...' command much closer to the current Python module.  The remaining
differences are the presence of a 'roots' indirection in the Python
output (currently used to pass through ref-engines-object-level
casEngines entries) and the lack of 'uri' entries for each root (which
I'm planning on adding).