tools/refengine/root: Clear MediaType and URI if unset in JSON
And add tests to confirm this.  I expect this is unlikely to matter in
practice, but I think it makes sense to have the unmarshalled JSON be
more like PUT [1] than PATCH [2].  For what it's worth, the Go
implementation seems to prefer PATCH (at least as of 1.9).

Also add similar tests to tools/engine/config_test.go, although we
didn't need changes to the implementation to get those to pass.

[1]: https://tools.ietf.org/html/rfc7231#section-4.3.4
[2]: https://tools.ietf.org/html/rfc5789