change `handleIndex` to `getMerkleRoots`
To remove http response stuff (checking mediatype and decoding body)
from index handle function to `*engine.Get`.

We should only keep this internal function to parse out the merkle
roots.

And improve the unit test to fit the changed function. It has to remove
the test cases for checking bad index json format. Because it exists no
more in new function. Maybe we should check it in `*engine.Get` tests.

Signed-off-by: xiekeyang <xiekeyang@huawei.com>