tools/indextemplate: Refactor into getPreFetch and getPostFetch
Go doesn't seem to have an analog to Python's unittest.mock.patch [1],
so increase test coverage by moving as much as possible out of Get and
into pre- and post-fetch helper methods.  With this commit, we have:

  $ go test -cover ./tools/indextemplate/...
  ok      github.com/xiekeyang/oci-discovery/tools/indextemplate  0.004s  coverage: 74.1% of statements

while before this commit we had:

  $ go test -cover ./tools/indextemplate/...
  ok      github.com/xiekeyang/oci-discovery/tools/indextemplate  0.003s  coverage: 33.9% of statements

although not all of that difference is due to better Get
decomposition.

Some of the tests added in this commit are being restored after
1ecc6df (Parse out Merkle Roots: change `handleIndex` to
`getMerkleRoots`, 2017-09-23, #40) removed them.

[1]: https://docs.python.org/3.5/library/unittest.mock.html#unittest.mock.patch